+++ apps/wallet/src/ui/app/redux/slices/transaction-requests/index.ts
@@ -3,6 +3,8 @@
 
 import {
     Base64DataBuffer,
+    getCertifiedTransaction,
+    getTransactionEffects,
     type SuiMoveNormalizedFunction,
 } from '@mysten/sui.js';
 import {
@@ -11,7 +13,13 @@ import {
     createSlice,
 } from '@reduxjs/toolkit';
 
-import type { SuiTransactionResponse } from '@mysten/sui.js';
+import { FEATURES } from '_src/ui/app/experimentation/features';
+
+import type {
+    SuiTransactionResponse,
+    SignableTransaction,
+    SuiExecuteTransactionResponse,
+} from '@mysten/sui.js';
 import type { PayloadAction } from '@reduxjs/toolkit';
 import type { TransactionRequest } from '_payloads/transactions';
 import type { RootState } from '_redux/RootReducer';
@@ -68,7 +76,7 @@ export const respondToTransactionRequest = createAsyncThunk<
     'respond-to-transaction-request',
     async (
         { txRequestID, approved },
-        { extra: { background, api, keypairVault }, getState }
+        { extra: { background, api, keypairVault, featureGating }, getState }
     ) => {
         const state = getState();
         const txRequest = txRequestsSelectors.selectById(state, txRequestID);
@@ -80,19 +88,64 @@ export const respondToTransactionRequest = createAsyncThunk<
         if (approved) {
             const signer = api.getSignerInstance(keypairVault.getKeyPair());
             try {
-                if (txRequest.tx.type === 'v2') {
-                    txResult = await signer.signAndExecuteTransaction(
-                        txRequest.tx.data
-                    );
-                } else if (txRequest.tx.type === 'move-call') {
-                    txResult = await signer.executeMoveCall(txRequest.tx.data);
-                } else if (txRequest.tx.type === 'serialized-move-call') {
-                    const txBytes = new Base64DataBuffer(txRequest.tx.data);
-                    txResult = await signer.signAndExecuteTransaction(txBytes);
+                if (featureGating.isOn(FEATURES.DEPRECATE_GATEWAY)) {
+                    let response: SuiExecuteTransactionResponse;
+                    if (
+                        txRequest.tx.type === 'v2' ||
+                        txRequest.tx.type === 'move-call'
+                    ) {
+                        const txn: SignableTransaction =
+                            txRequest.tx.type === 'move-call'
+                                ? {
+                                      kind: 'moveCall',
+                                      data: txRequest.tx.data,
+                                  }
+                                : txRequest.tx.data;
+
+                        response =
+                            await signer.signAndExecuteTransactionWithRequestType(
+                                txn
+                            );
+                    } else if (txRequest.tx.type === 'serialized-move-call') {
+                        const txBytes = new Base64DataBuffer(txRequest.tx.data);
+                        response =
+                            await signer.signAndExecuteTransactionWithRequestType(
+                                txBytes
+                            );
+                    } else {
+                        throw new Error(
+                            `Either tx or txBytes needs to be defined.`
+                        );
+                    }
+
+                    txResult = {
+                        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                        certificate: getCertifiedTransaction(response)!,
+                        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                        effects: getTransactionEffects(response)!,
+                        timestamp_ms: null,
+                        parsed_data: null,
+                    };
                 } else {
-                    throw new Error(
-                        `Either tx or txBytes needs to be defined.`
-                    );
+                    await signer.syncAccountState();
+                    if (txRequest.tx.type === 'v2') {
+                        txResult = await signer.signAndExecuteTransaction(
+                            txRequest.tx.data
+                        );
+                    } else if (txRequest.tx.type === 'move-call') {
+                        txResult = await signer.executeMoveCall(
+                            txRequest.tx.data
+                        );
+                    } else if (txRequest.tx.type === 'serialized-move-call') {
+                        const txBytes = new Base64DataBuffer(txRequest.tx.data);
+                        txResult = await signer.signAndExecuteTransaction(
+                            txBytes
+                        );
+                    } else {
+                        throw new Error(
+                            `Either tx or txBytes needs to be defined.`
+                        );
+                    }
                 }
             } catch (e) {
                 tsResultError = (e as Error).message;