+++ sdk/typescript/src/signers/signer-with-provider.ts
@@ -62,6 +62,9 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This method will be removed soon after we deprecate gateway. Prefer to use
+   * `signAndExecuteTransactionWithRequestType`
+   *
    * Sign a transaction and submit to the Gateway for execution
    */
   async signAndExecuteTransaction(
@@ -99,6 +102,8 @@ export abstract class SignerWithProvider implements Signer {
         return this.splitCoin(transaction.data);
       case 'pay':
         return this.pay(transaction.data);
+      case 'publish':
+        return this.publish(transaction.data);
       default:
         throw new Error(
           `Unknown transaction kind: "${(transaction as any).kind}"`
@@ -107,11 +112,12 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
-   * @experimental Sign a transaction and submit to the Fullnode for execution
+   * Sign a transaction and submit to the Fullnode for execution. Only exists
+   * on Fullnode
    */
   async signAndExecuteTransactionWithRequestType(
     transaction: Base64DataBuffer | SignableTransaction,
-    requestType: ExecuteTransactionRequestType
+    requestType: ExecuteTransactionRequestType = 'WaitForLocalExecution'
   ): Promise<SuiExecuteTransactionResponse> {
     // Handle submitting raw transaction bytes:
     if (
@@ -152,6 +158,8 @@ export abstract class SignerWithProvider implements Signer {
         return this.splitCoinWithRequestType(transaction.data, requestType);
       case 'pay':
         return this.payWithRequestType(transaction.data, requestType);
+      case 'publish':
+        return this.publishWithRequestType(transaction.data, requestType);
       default:
         throw new Error(
           `Unknown transaction kind: "${(transaction as any).kind}"`
@@ -160,6 +168,7 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This API will be removed soon after we deprecate gateway
    * Trigger gateway to sync account state related to the address,
    * based on the account state on validators.
    */
@@ -169,6 +178,9 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This API will be removed soon after we deprecate gateway. Prefer to use
+   * `signAndExecuteTransactionWithRequestType`
+   *
    * Serialize and Sign a `TransferObject` transaction and submit to the Gateway for execution
    */
   async transferObject(
@@ -183,6 +195,9 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This method will be removed soon after we deprecate gateway. Prefer to use
+   * `transferSuiWithRequestType`
+   *
    * Serialize and Sign a `TransferSui` transaction and submit to the Gateway for execution
    */
   async transferSui(
@@ -197,6 +212,9 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This method will be removed soon after we deprecate gateway. Prefer to use
+   * `payWithRequestType`
+   *
    * Serialize and Sign a `Pay` transaction and submit to the Gateway for execution
    */
   async pay(transaction: PayTransaction): Promise<SuiTransactionResponse> {
@@ -206,6 +224,9 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This method will be removed soon after we deprecate gateway. Prefer to use
+   * `mergeCoinWithRequestType`
+   *
    * Serialize and Sign a `MergeCoin` transaction and submit to the Gateway for execution
    */
   async mergeCoin(
@@ -220,6 +241,9 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This method will be removed soon after we deprecate gateway. Prefer to use
+   * `splitCoinWithRequestType`
+   *
    * Serialize and Sign a `SplitCoin` transaction and submit to the Gateway for execution
    */
   async splitCoin(
@@ -234,6 +258,9 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This method will be removed soon after we deprecate gateway. Prefer to use
+   * `executeMoveCallWithRequestType`
+   *
    * Serialize and Sign a `MoveCall` transaction and submit to the Gateway for execution
    */
   async executeMoveCall(
@@ -248,6 +275,9 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
+   * @deprecated This method will be removed soon after we deprecate gateway. Prefer to use
+   * `publishWithRequestType`
+   *
    * Publish a Move package on chain
    * @param transaction See {@link PublishTransaction}
    */
@@ -262,9 +292,9 @@ export abstract class SignerWithProvider implements Signer {
     return await this.signAndExecuteTransaction(txBytes);
   }
 
-  /* ---------------------------- Experimental API ---------------------------- */
   /**
-   * @experimental Serialize and sign a `TransferObject` transaction and submit to the Fullnode
+   *
+   * Serialize and sign a `TransferObject` transaction and submit to the Fullnode
    * for execution
    */
   async transferObjectWithRequestType(
@@ -283,7 +313,8 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
-   * @experimental Serialize and sign a `TransferSui` transaction and submit to the Fullnode
+   *
+   * Serialize and sign a `TransferSui` transaction and submit to the Fullnode
    * for execution
    */
   async transferSuiWithRequestType(
@@ -302,7 +333,8 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
-   * @experimental Serialize and Sign a `Pay` transaction and submit to the fullnode for execution
+   *
+   * Serialize and Sign a `Pay` transaction and submit to the fullnode for execution
    */
   async payWithRequestType(
     transaction: PayTransaction,
@@ -317,7 +349,8 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
-   * @experimental Serialize and sign a `MergeCoin` transaction and submit to the Fullnode
+   *
+   * Serialize and sign a `MergeCoin` transaction and submit to the Fullnode
    * for execution
    */
   async mergeCoinWithRequestType(
@@ -336,7 +369,8 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
-   * @experimental Serialize and sign a `SplitCoin` transaction and submit to the Fullnode
+   *
+   * Serialize and sign a `SplitCoin` transaction and submit to the Fullnode
    * for execution
    */
   async splitCoinWithRequestType(
@@ -355,7 +389,7 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
-   * @experimental Serialize and sign a `MoveCall` transaction and submit to the Fullnode
+   * Serialize and sign a `MoveCall` transaction and submit to the Fullnode
    * for execution
    */
   async executeMoveCallWithRequestType(
@@ -374,7 +408,8 @@ export abstract class SignerWithProvider implements Signer {
   }
 
   /**
-   * @experimental Serialize and sign a `Publish` transaction and submit to the Fullnode
+   *
+   * Serialize and sign a `Publish` transaction and submit to the Fullnode
    * for execution
    */
   async publishWithRequestType(