+++ sdk/wallet-adapter/packages/adapters/all-wallets/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mysten/wallet-adapter-all-wallets
 
+## 0.2.1
+
+### Patch Changes
+
+- @mysten/wallet-adapter-mock-wallet@0.2.1
+- @mysten/wallet-adapter-sui-wallet@0.2.1
+- @mysten/wallet-adapter-wallet-standard@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes